camelcase filenames and other naming changes
