+++ webscraping/courseInfo/storing/courseInfoSchema.json
