+++ webscraping/courseInfo/storing/storer.js
@@ -1,31 +1,32 @@
-'use strict';
+'use strict'
 
-const nodemailer = require('nodemailer');
-const remove = require("remove");
-const fs = require("fs");
-const MongoClient = require('mongodb').MongoClient;
-const assert = require('assert');
-const argv = require('minimist')(process.argv.slice(2));
-const Ajv = require('ajv');
-const ajv = new Ajv({
-    "verbose": true,
-    "allErrors": true
-});
+let nodemailer = require('nodemailer')
+let remove = require("remove")
+let fs = require("fs")
+let MongoClient = require('mongodb').MongoClient
+let assert = require('assert')
+let argv = require('minimist')(process.argv.slice(2))
 
-const validate = ajv.compile(JSON.parse(fs.readFileSync('recommendedSequence.json', 'utf8')));
+const json_schema_path = 'courseInfoSchema.json'
+const scraped_course_info_folder = '../scrapedJson'
+const course_info_document_regex = /_document/
+
+let Ajv = require('ajv');
+let ajv = new Ajv({ "verbose": true, "allErrors": true });
+let validate = ajv.compile(JSON.parse(fs.readFileSync(json_schema_path, 'utf8')));
+
 const mongoServerUrl = 'mongodb://138.197.6.26:27017/';
 const devDbName = "courseplannerdb-dev";
 const prodDbName = "courseplannerdb";
 const dbName = (argv.prod) ? prodDbName : devDbName;
 const dbFullUrl = mongoServerUrl + dbName;
 
-let log = "*** Sequence Validation Log ***<br><br>";
+let log = "*** Course Info Validation Log ***<br><br>";
 
-const storeAllSequences = (function (){
+const storeAllCourseInfo = (function (){
 
     console.log("Storing + validating sequence json data");
 
-    let seqFolder = '../scraped-json/';
     let numValidated = 0;
 
     MongoClient.connect(dbFullUrl, function(err, db) {
@@ -34,16 +35,17 @@ const storeAllSequences = (function (){
 
         let foundIssue = false;
 
-        // read all files in sequence folder and pass them through the validator
-        fs.readdir(seqFolder, function (err, files) {
+        // read all files in course info folder and pass them through the validator
+        fs.readdir(scraped_course_info_folder, function (err, files) {
+            files = files.filter(file => course_info_document_regex.test(file))
             files.forEach(function (file) {
-                fs.readFile(seqFolder + file, "utf-8", function (err, fileContent) {
-                    let sequenceJSON = JSON.parse(fileContent);
+                fs.readFile(scraped_course_info_folder + '/' + file, "utf-8", function (err, fileContent) {
                     if (err) {
                         throw err;
                     }
+                    let courseInfosJSON = JSON.parse(fileContent);
 
-                    let isSequenceValid = validate(sequenceJSON);
+                    let isSequenceValid = validate(courseInfosJSON);
 
                     numValidated++;
                     if(!isSequenceValid){
@@ -54,10 +56,12 @@ const storeAllSequences = (function (){
                         logMessage(file + ": PASS");
 
                         // write the json to the db
-                        db.collection("courseSequences").update({_id : file.replace(".json","")}, {$set:sequenceJSON}, {upsert: true}, function(err, result) {
-                            assert.equal(err, null);
-                            logMessage("Wrote contents of file: " + file + " to db.");
-                        });
+                        courseInfosJSON.forEach(courseInfoJSON => {
+                            db.collection("courseData").update({_id : courseInfoJSON.code}, {$set:courseInfoJSON}, {upsert: true}, function(err, result) {
+                                assert.equal(err, null)
+                                logMessage("Wrote contents of file: " + courseInfoJSON.code + " to db.")
+                            })
+                        })
                     }
 
                     if (numValidated == files.length) {
@@ -72,15 +76,6 @@ const storeAllSequences = (function (){
     });
 })();
 
-function cleanUp(){
-    try {
-        remove.removeSync("../scraped-json");
-        logMessage("cleaned up sequences dir");
-    } catch (err) {
-        console.error(err);
-    }
-}
-
 function logMessage(message){
     console.log(message);
     log += message + "<br>"
@@ -88,7 +83,7 @@ function logMessage(message){
 
 function sendIssueEmail(){
 
-    let message = "The course sequence scraper encountered errors in its most recent execution (" + new Date().toString() + ")\n" +
+    let message = "The course info scraper encountered errors in its most recent execution (" + new Date().toString() + ")\n" +
                     " Below are the logs from the scrape attempt:<br><br>";
     message += log;
