+++ webscraping/courseSequences/storing/courseSequenceSchema.json
